+++ src/commands/packsearch.js
@@ -105,9 +105,9 @@ module.exports = {
 				if (i.customId === 'upload') {
 					await i.update({ embeds: [embeds[index]], components: [pageButtons, actionButtonsDisabled] });
 
-					if (!interaction.member.permissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS)) {
+					if (!i.memberPermissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS)) {
 						return interaction.editReply({
-							content: 'You do not have enough permissions to use this command.\nYou need Manage Emojis perms to use this command.',
+							content: 'You do not have enough permissions to use this command.\nRequires **Manage Emojis**.',
 							ephemeral: true,
 						});
 					}