+++ src/commands/random.js
@@ -61,7 +61,7 @@ module.exports = {
 				if (i.customId === 'upload') {
 					await i.update({ embeds: [embed], components: [disabledRow] });
 
-					if (!interaction.member.permissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS)) {
+					if (!i.memberPermissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS)) {
 						return interaction.editReply({
 							content: 'You do not have enough permissions to use this command.\nRequires **Manage Emojis**.',
 							ephemeral: true,