+++ src/commands/removeunused.js
@@ -66,7 +66,7 @@ module.exports = {
 				if (i.customId === 'remove') {
 					await i.update({ components: [] });
 
-					if (!interaction.member.permissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS)) {
+					if (!i.memberPermissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS)) {
 						return interaction.editReply({
 							content: 'You do not have enough permissions to use this command.\nRequires **Manage Emojis**.',
 							ephemeral: true,